Diet service

+++ paprika-service/src/main/kotlin/com/rmp/paprika/dto/dish/DishDto.kt
@@ -4,13 +4,15 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class DishDto (
-    val id: Int,
+    val id: Long,
     val name: String,
     val logo: String,
+
     val calories: Double,
     val protein: Double,
     val fat: Double,
     val carbohydrates: Double,
+
     val timeToCook: Int,
-    val typeId: Int,
+    val typeId: Long,
 )
